chore: update one more place for time.Duraiton

+++ cmd/influx/setup.go
@@ -240,7 +240,7 @@ func nonInteractive() (*influxdb.OnboardingRequest, error) {
 		return nil, err
 	}
 	if dur > 0 {
-		req.RetentionPeriod = uint(dur / time.Hour)
+		req.RetentionPeriod = dur / time.Hour
 	}
 	return req, nil
 }
